Format with new files
