+++ codalab/worker/main.py
@@ -17,7 +17,7 @@ import multiprocessing
 from codalab.lib.formatting import parse_size
 from .bundle_service_client import BundleServiceClient, BundleAuthException
 from . import docker_utils
-from .worker import Worker
+from worker import Worker
 from dependency_manager import DependencyManager
 from docker_image_manager import DockerImageManager
 from run_manager import RunManager
